Fix deadlock caused by SortedSet (#4)
Turns out Sorted Set was causing an issue when we received the same offset twice. The second occurrence didn't get recorded because sorted set doesn't allow duplicated. We also saw an issue trying to keep two arrays (received_offsets and completed_offsets) properly reconciled at high volume and was not able to trace down the cause. We moved to a single array (received_offsets) for tracking with the addition of OffsetHolder which holds the state of each offset. This array is sorted by partition/offset and completed offsets are ack'd and removed from the stack.

Bump version to 0.1.3.1.